removed line that tried to see if user was signing up multiple times
in my tests it seems like this doesn't even work in Orthus's current
Google OAuth implementation. If a user tries to sign up again with the
same account it just replaces the information filled out with the new
entry.

It might be a good idea to prevent multiple sign ups in the future, but
this doesn't do the job.